Include migrations in wheels (#686)
* Include migrations in wheels

Whoops ...

* Don't use editable install for CI

* Remove aequilbrae dir after install

Bit of a hack to avoid the relative import

* improvements

* improvements

* Don't rm aequilibrae with renamed dir

* Go back to editable install, too much work to fix properly

* .

* Lint

* .

---------
